feature: Multicast nodes discovery for `ants::Worker`.
feature: Added self test for multicast agent.

BREAKING CHANGE: `ants::Worker::new_arc` changed to an async equivalent
of `ants::Worker::new_and_init` which includes starting the broadcaster.

feature: added `NodeList` composite in order for future broadcasts to
effectively add and remove nodes. `NodeList::checkout` will handle the re-insertion of
`NodeAddress` after attempted reservation.

+++ packages/multicast/src/proto/implementations/hash.rs
@@ -41,7 +41,7 @@ impl Hash for MulticastMessage {
     fn hash<H: Hasher>(&self, state: &mut H) {
         hash_multicast_message_components(
             &self.uuid,
-            &(self.kind as i32),
+            &self.kind,
             self.timestamp.as_ref(),
             &self.body,
             &self.ttl,
@@ -56,7 +56,7 @@ impl MulticastMessage {
     pub fn generate_checksum(&self) -> u64 {
         multicast_message_components_checksum(
             &self.uuid,
-            &(self.kind as i32),
+            &self.kind,
             self.timestamp.as_ref(),
             &self.body,
             &self.ttl,
